// src/api/generated/model/eventTaskDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */
import type { EventTaskDtoAllowedFormatsItem } from './eventTaskDtoAllowedFormatsItem';
import type { EventTaskDtoTargetPlatform } from './eventTaskDtoTargetPlatform';
import type { EventTaskDtoOrdForm } from './eventTaskDtoOrdForm';
import type { EventTaskDtoOrdFlagsItem } from './eventTaskDtoOrdFlagsItem';
import type { EventTaskDtoOrdPayType } from './eventTaskDtoOrdPayType';
import type { EventTaskDtoEventType } from './eventTaskDtoEventType';
import type { EventTaskDtoReward } from './eventTaskDtoReward';

export interface EventTaskDto {
  /** Unique identifier */
  id: string;
  /** Timestamp of when the record was created */
  createdAt: string;
  /** Timestamp of the last update of the record */
  updatedAt: string;
  /** Название задания */
  title: string;
  /**
   * Описание задания
   * @nullable
   */
  description?: string | null;
  /** Флаг удаления */
  isDeleted: boolean;
  /** Заморожено ли задание владельцем комнаты */
  isFrozen: boolean;
  /** Критерии выполнения задания */
  criteria?: string[];
  /** Что запрещено при выполнении задания */
  restrictions?: string[];
  /** Допустимые форматы ответа на задание */
  allowedFormats?: EventTaskDtoAllowedFormatsItem[];
  /** Целевая платформа публикации */
  targetPlatform: EventTaskDtoTargetPlatform;
  /**
   * Форма распространения креатива (тип креатива)
   * @nullable
   */
  ordForm?: EventTaskDtoOrdForm;
  /** Особый тип рекламного объявления (вид рекламы). native на запись не поддерживается ОРД */
  ordFlags?: EventTaskDtoOrdFlagsItem[];
  /** Коды ККТУ (1 для обычного креатива, до 16 для кобрендингового) */
  ordKktus?: string[];
  /**
   * Бренд рекламируемых товаров или услуг
   * @nullable
   */
  ordBrand?: string | null;
  /**
   * Вид рекламируемых товаров или услуг
   * @nullable
   */
  ordCategory?: string | null;
  /**
   * Дополнительное описание рекламируемых товаров или услуг
   * @nullable
   */
  ordProductDescription?: string | null;
  /**
   * Описание целевой аудитории
   * @nullable
   */
  ordTargeting?: string | null;
  /**
   * Тип оплаты креатива (pay_type в ОРД)
   * @nullable
   */
  ordPayType?: EventTaskDtoOrdPayType;
  /**
   * Точное количество публикаций (под-ответов), которое обязан подготовить амбассадор
   * @minimum 1
   * @maximum 100
   */
  publicationsCount: number;
  /** Требуется ли этап проверки материалов перед публикацией. false — этап пропускается. */
  requireMaterialsReview: boolean;
  /** Требуется ли этап проверки опубликованного контента перед начислением награды. false — после отчёта о публикации ответ сразу одобряется. */
  requirePublicationReview: boolean;
  /** @nullable */
  ordContractTemplateId: string | null;
  eventId: string;
  eventType: EventTaskDtoEventType;
  experiencePoints: number;
  /** @nullable */
  reward: EventTaskDtoReward;
}